FER-2844 Moving/renaming classes according to Ethan's comments.
