+++ flowz/channels/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
 
-from .channels import (ChannelDone, Channel, ReadChannel, MapChannel, FlatMapChannel,
-                       FilterChannel, FutureChannel, ReadyFutureChannel, TeeChannel,
-                       ProducerChannel, IterChannel, ZipChannel, CoGroupChannel)
+from .core import (ChannelDone, Channel, ReadChannel, MapChannel, FlatMapChannel,
+                   FilterChannel, FutureChannel, ReadyFutureChannel, TeeChannel,
+                   ProducerChannel, IterChannel, ZipChannel, CoGroupChannel)
